Atualização

+++ dscommerce-secao3/src/main/java/com/walker/dscommerce/service/ProductService.java
@@ -1,12 +1,17 @@
 package com.walker.dscommerce.service;
 
 import com.walker.dscommerce.dto.ProductDTO;
+import com.walker.dscommerce.exception.DataBaseIntegrityViolationException;
+import com.walker.dscommerce.exception.ResourceNotFoundException;
 import com.walker.dscommerce.model.Product;
 import com.walker.dscommerce.repository.ProductRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.modelmapper.ModelMapper;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 
@@ -27,7 +32,7 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
         Product product = productRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Product não encontrado com ID: " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Product não encontrado com ID: " + id)); //Tratando e lançando a exceção customizada
         return modelMapper.map(product, ProductDTO.class);
     }
 
@@ -47,18 +52,28 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product product = productRepository.getReferenceById(id);
-        product.setName(productDTO.getName());
-        product.setDescription(productDTO.getDescription());
-        product.setPrice(productDTO.getPrice());
-        product.setImgUrl(productDTO.getImgUrl());
-        productRepository.save(product);
-        return modelMapper.map(product, ProductDTO.class);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            product.setName(productDTO.getName());
+            product.setDescription(productDTO.getDescription());
+            product.setPrice(productDTO.getPrice());
+            product.setImgUrl(productDTO.getImgUrl());
+            productRepository.save(product);
+            return modelMapper.map(product, ProductDTO.class);
+        } catch (EntityNotFoundException entityNotFoundException) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        productRepository.findById(id).orElseThrow(() -> new RuntimeException("Product não encontrado com ID: " + id));
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Product não encontrado com ID: " + id);
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException dataIntegrityViolationException) {
+            throw new DataBaseIntegrityViolationException("Falha de integridade referencial!");
+        }
     }
 }
